refactor: extract SummaryTabNavigation
