--- src/app/(main)/summary/_components/SummaryResultCard.tsx
@@ -1,91 +1,0 @@
-'use client'
-
-import { useEffect } from 'react'
-import Image from 'next/image'
-import { useRouter } from 'next/navigation'
-
-import { Again } from '@/components/icon'
-import { useSummaryQuery } from '@/hooks/fetch/useSummaryQuery'
-import { SummaryForm, addSummary } from '@/services/summary'
-import { Button } from '@/components/ui/Button'
-import { SummaryResultHeader } from './summaryResultCard/SummaryResultHeader'
-
-const activeBtnStyle = 'bg-black text-white'
-
-const summaryTypes = [
-  { key: 'basic', label: '기본 요약' },
-  { key: 'oneline', label: '한 줄 요약' },
-  { key: 'fiveline', label: '다섯 줄 요약' },
-  { key: 'insight', label: '인사이트' },
-] as const
-
-export function SummaryResultCard() {
-  const router = useRouter()
-  const { data: summaryData, isLoading } = useSummaryQuery()
-
-  const { status, subject, content, text, type, origin_url: originUrl } = summaryData?.data?.data ?? {}
-
-  useEffect(() => {
-    if (status === 'failed') {
-      alert('요약에 실패하였습니다. 다시 시도해주세요.')
-      router.push(`/summary`)
-    }
-  }, [status, router])
-
-  if (!summaryData) return null
-
-  const onSubmit = async (data: SummaryForm) => {
-    const response = await addSummary(data)
-
-    if (response.status === 202) {
-      router.push(`/summary/${response.data.data.task_id}`)
-    } else {
-      alert('요약을 실패하였습니다. 다시 시도해주세요.')
-    }
-  }
-
-  return (
-    <>
-      {status === 'completed' || isLoading ? (
-        <>
-          <SummaryResultHeader />
-          <hr className="border-gray-light -mx-10 h-px" />
-          <div className="flex-1 overflow-auto">
-            <div className="mx-auto mb-5 max-w-2xl text-3xl font-semibold">
-              <div className="mt-10 leading-[1.5]">{subject}</div>
-            </div>
-            <div className="flex justify-between">
-              <div>
-                {summaryTypes.map(({ key, label }) => (
-                  <Button
-                    key={key}
-                    onClick={() => onSubmit({ url: originUrl ?? '', content: text ?? '', type: key })}
-                    variant={'SUMMARY_CATEGORY_BUTTON'}
-                    className={type === key ? activeBtnStyle : ''}
-                  >
-                    {label}
-                  </Button>
-                ))}
-              </div>
-              <div className="inline-flex rounded-full bg-[linear-gradient(90deg,#BEBDFF_0%,#9F6BC4_48%,#E26466_100%)] p-[1px]">
-                <button
-                  className="flex cursor-pointer items-center gap-1 rounded-full bg-white px-4 py-2 text-sm text-black"
-                  onClick={() => onSubmit({ url: originUrl ?? '', content: text ?? '', type: type ?? '' })}
-                >
-                  재요약
-                  <Again />
-                </button>
-              </div>
-            </div>
-            <hr className="border-gray-light mb-9 mt-4 border" />
-            <div>{content}</div>
-          </div>
-        </>
-      ) : (
-        <div className="flex h-full items-center justify-center">
-          <Image src={'/assets/loading.gif'} alt="loading" width={100} height={100} />
-        </div>
-      )}
-    </>
-  )
-}
