+++ src/app/(main)/summary/_components/summaryResultCard/index.ts
@@ -1,1 +1,4 @@
+export { SummaryResultCard } from './SummaryResultCard'
+
 export { SummaryResultHeader } from './SummaryResultHeader'
+export { SummaryTabNavigation } from './SummaryTabNavigation'
